requirements.txt
# Google ADK
google-adk[a2a,eval]>=1.18.0

# RSS Feed Parsing
feedparser==6.0.10

# Social Media APIs
tweepy==4.14.0              # Twitter
requests-oauthlib==1.3.1    # OAuth for LinkedIn
requests>=2.32.4            # HTTP requests

# Database
aiosqlite==0.19.0           # Async SQLite
sqlalchemy==2.0.23          # ORM (optional)

# Retry Logic
tenacity>=9.0.0             # Retry decorators

# Configuration
python-dotenv==1.0.0        # Environment variables

# Date/Time handling
pytz==2024.1                # Timezone support

# Utilities
opentelemetry-instrumentation-google-genai
